backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart>=0.0.7
SpeechRecognition==3.10.0
moviepy==1.0.3
pydub==0.25.1
yt-dlp>=2024.08.06
youtube-transcript-api==0.6.2
imageio-ffmpeg==0.4.9
brotli>=1.0.9
